requirements/base.txt
xrpl-py
glom
